+++ pkg/apis/sources/v1alpha2/container_conversion_test.go
@@ -21,8 +21,6 @@ import (
 	"reflect"
 	"testing"
 
-	"knative.dev/eventing/pkg/apis/sources"
-
 	corev1 "k8s.io/api/core/v1"
 
 	"github.com/google/go-cmp/cmp"
@@ -33,7 +31,7 @@ import (
 )
 
 func TestContainerSourceConversionBadType(t *testing.T) {
-	good, bad := &ContainerSource{}, &sources.DummyObject{}
+	good, bad := &ContainerSource{}, &dummyObject{}
 
 	if err := good.ConvertTo(context.Background(), bad); err == nil {
 		t.Errorf("ConvertTo() = %#v, wanted error", bad)
