+++ pkg/apis/sources/v1alpha2/container_lifecycle_test.go
@@ -19,8 +19,6 @@ package v1alpha2
 import (
 	"testing"
 
-	"knative.dev/eventing/pkg/apis/sources"
-
 	"github.com/google/go-cmp/cmp"
 	"github.com/google/go-cmp/cmp/cmpopts"
 	appsv1 "k8s.io/api/apps/v1"
@@ -91,7 +89,7 @@ func TestContainerSourceStatusIsReady(t *testing.T) {
 		s: func() *ContainerSourceStatus {
 			s := &ContainerSourceStatus{}
 			s.InitializeConditions()
-			s.PropagateReceiveAdapterStatus(sources.TestHelper.AvailableDeployment())
+			s.PropagateReceiveAdapterStatus(availableDeployment)
 			return s
 		}(),
 		wantConditionStatus: corev1.ConditionUnknown,
@@ -112,7 +110,7 @@ func TestContainerSourceStatusIsReady(t *testing.T) {
 			s := &ContainerSourceStatus{}
 			s.InitializeConditions()
 			s.PropagateSinkBindingStatus(&readySinkBinding.Status)
-			s.PropagateReceiveAdapterStatus(sources.TestHelper.AvailableDeployment())
+			s.PropagateReceiveAdapterStatus(availableDeployment)
 			return s
 		}(),
 		wantConditionStatus: corev1.ConditionTrue,
@@ -123,7 +121,7 @@ func TestContainerSourceStatusIsReady(t *testing.T) {
 			s := &ContainerSourceStatus{}
 			s.InitializeConditions()
 			s.PropagateSinkBindingStatus(&readySinkBinding.Status)
-			s.PropagateReceiveAdapterStatus(sources.TestHelper.UnavailableDeployment())
+			s.PropagateReceiveAdapterStatus(unavailableDeployment)
 			return s
 		}(),
 		wantConditionStatus: corev1.ConditionFalse,
@@ -134,7 +132,7 @@ func TestContainerSourceStatusIsReady(t *testing.T) {
 			s := &ContainerSourceStatus{}
 			s.InitializeConditions()
 			s.PropagateSinkBindingStatus(&readySinkBinding.Status)
-			s.PropagateReceiveAdapterStatus(sources.TestHelper.UnknownDeployment())
+			s.PropagateReceiveAdapterStatus(unknownDeployment)
 			return s
 		}(),
 		wantConditionStatus: corev1.ConditionUnknown,
@@ -156,7 +154,7 @@ func TestContainerSourceStatusIsReady(t *testing.T) {
 			s := &ContainerSourceStatus{}
 			s.InitializeConditions()
 			s.PropagateSinkBindingStatus(&readySinkBinding.Status)
-			s.PropagateReceiveAdapterStatus(sources.TestHelper.AvailableDeployment())
+			s.PropagateReceiveAdapterStatus(availableDeployment)
 			s.PropagateSinkBindingStatus(&notReadySinkBinding.Status)
 			return s
 		}(),
@@ -167,9 +165,9 @@ func TestContainerSourceStatusIsReady(t *testing.T) {
 		s: func() *ContainerSourceStatus {
 			s := &ContainerSourceStatus{}
 			s.InitializeConditions()
-			s.PropagateReceiveAdapterStatus(sources.TestHelper.AvailableDeployment())
+			s.PropagateReceiveAdapterStatus(availableDeployment)
 			s.PropagateSinkBindingStatus(&notReadySinkBinding.Status)
-			s.PropagateReceiveAdapterStatus(sources.TestHelper.UnavailableDeployment())
+			s.PropagateReceiveAdapterStatus(unavailableDeployment)
 			return s
 		}(),
 		wantConditionStatus: corev1.ConditionFalse,
@@ -180,7 +178,7 @@ func TestContainerSourceStatusIsReady(t *testing.T) {
 			s := &ContainerSourceStatus{}
 			s.InitializeConditions()
 			s.PropagateSinkBindingStatus(&notReadySinkBinding.Status)
-			s.PropagateReceiveAdapterStatus(sources.TestHelper.AvailableDeployment())
+			s.PropagateReceiveAdapterStatus(availableDeployment)
 			return s
 		}(),
 		wantConditionStatus: corev1.ConditionFalse,
@@ -191,7 +189,7 @@ func TestContainerSourceStatusIsReady(t *testing.T) {
 			s := &ContainerSourceStatus{}
 			s.InitializeConditions()
 			s.PropagateSinkBindingStatus(&notReadySinkBinding.Status)
-			s.PropagateReceiveAdapterStatus(sources.TestHelper.AvailableDeployment())
+			s.PropagateReceiveAdapterStatus(availableDeployment)
 			s.PropagateSinkBindingStatus(&readySinkBinding.Status)
 			return s
 		}(),
@@ -243,7 +241,7 @@ func TestContainerSourceStatusGetCondition(t *testing.T) {
 		s: func() *ContainerSourceStatus {
 			s := &ContainerSourceStatus{}
 			s.InitializeConditions()
-			s.PropagateReceiveAdapterStatus(sources.TestHelper.AvailableDeployment())
+			s.PropagateReceiveAdapterStatus(availableDeployment)
 			return s
 		}(),
 		condQuery: ContainerSourceConditionReady,
@@ -270,7 +268,7 @@ func TestContainerSourceStatusGetCondition(t *testing.T) {
 			s := &ContainerSourceStatus{}
 			s.InitializeConditions()
 			s.PropagateSinkBindingStatus(&readySinkBinding.Status)
-			s.PropagateReceiveAdapterStatus(sources.TestHelper.AvailableDeployment())
+			s.PropagateReceiveAdapterStatus(availableDeployment)
 			return s
 		}(),
 		condQuery: ContainerSourceConditionReady,
@@ -284,7 +282,7 @@ func TestContainerSourceStatusGetCondition(t *testing.T) {
 			s := &ContainerSourceStatus{}
 			s.InitializeConditions()
 			s.PropagateSinkBindingStatus(&readySinkBinding.Status)
-			s.PropagateReceiveAdapterStatus(sources.TestHelper.AvailableDeployment())
+			s.PropagateReceiveAdapterStatus(availableDeployment)
 			s.PropagateSinkBindingStatus(&notReadySinkBinding.Status)
 			return s
 		}(),
@@ -301,8 +299,8 @@ func TestContainerSourceStatusGetCondition(t *testing.T) {
 			s := &ContainerSourceStatus{}
 			s.InitializeConditions()
 			s.PropagateSinkBindingStatus(&readySinkBinding.Status)
-			s.PropagateReceiveAdapterStatus(sources.TestHelper.AvailableDeployment())
-			s.PropagateReceiveAdapterStatus(sources.TestHelper.UnavailableDeployment())
+			s.PropagateReceiveAdapterStatus(availableDeployment)
+			s.PropagateReceiveAdapterStatus(unavailableDeployment)
 			return s
 		}(),
 		condQuery: ContainerSourceConditionReady,
@@ -316,7 +314,7 @@ func TestContainerSourceStatusGetCondition(t *testing.T) {
 			s := &ContainerSourceStatus{}
 			s.InitializeConditions()
 			s.PropagateSinkBindingStatus(&notReadySinkBinding.Status)
-			s.PropagateReceiveAdapterStatus(sources.TestHelper.AvailableDeployment())
+			s.PropagateReceiveAdapterStatus(availableDeployment)
 			s.PropagateSinkBindingStatus(&readySinkBinding.Status)
 			return s
 		}(),
