+++ pkg/apis/sources/v1alpha2/ping_lifecycle_test.go
@@ -19,8 +19,6 @@ package v1alpha2
 import (
 	"testing"
 
-	"knative.dev/eventing/pkg/apis/sources"
-
 	appsv1 "k8s.io/api/apps/v1"
 	corev1 "k8s.io/api/core/v1"
 
@@ -80,7 +78,7 @@ func TestPingSourceStatusIsReady(t *testing.T) {
 		s: func() *PingSourceStatus {
 			s := &PingSourceStatus{}
 			s.InitializeConditions()
-			s.PropagateDeploymentAvailability(sources.TestHelper.AvailableDeployment())
+			s.PropagateDeploymentAvailability(availableDeployment)
 			return s
 		}(),
 		wantConditionStatus: corev1.ConditionUnknown,
@@ -122,7 +120,7 @@ func TestPingSourceStatusIsReady(t *testing.T) {
 			s := &PingSourceStatus{}
 			s.InitializeConditions()
 			s.MarkSink(exampleUri)
-			s.PropagateDeploymentAvailability(sources.TestHelper.AvailableDeployment())
+			s.PropagateDeploymentAvailability(availableDeployment)
 			return s
 		}(),
 		wantConditionStatus: corev1.ConditionUnknown,
@@ -145,7 +143,7 @@ func TestPingSourceStatusIsReady(t *testing.T) {
 			s.InitializeConditions()
 			s.MarkSchedule()
 			s.MarkSink(exampleUri)
-			s.PropagateDeploymentAvailability(sources.TestHelper.AvailableDeployment())
+			s.PropagateDeploymentAvailability(availableDeployment)
 			return s
 		}(),
 		wantConditionStatus: corev1.ConditionTrue,
@@ -157,7 +155,7 @@ func TestPingSourceStatusIsReady(t *testing.T) {
 			s.InitializeConditions()
 			s.MarkSchedule()
 			s.MarkSink(exampleUri)
-			s.PropagateDeploymentAvailability(sources.TestHelper.UnavailableDeployment())
+			s.PropagateDeploymentAvailability(unavailableDeployment)
 			return s
 		}(),
 		wantConditionStatus: corev1.ConditionFalse,
@@ -169,7 +167,7 @@ func TestPingSourceStatusIsReady(t *testing.T) {
 			s.InitializeConditions()
 			s.MarkSchedule()
 			s.MarkSink(exampleUri)
-			s.PropagateDeploymentAvailability(sources.TestHelper.UnknownDeployment())
+			s.PropagateDeploymentAvailability(unknownDeployment)
 			return s
 		}(),
 		wantConditionStatus: corev1.ConditionUnknown,
@@ -181,7 +179,7 @@ func TestPingSourceStatusIsReady(t *testing.T) {
 			s.InitializeConditions()
 			s.MarkSchedule()
 			s.MarkSink(exampleUri)
-			s.PropagateDeploymentAvailability(sources.TestHelper.AvailableDeployment())
+			s.PropagateDeploymentAvailability(availableDeployment)
 			s.MarkEventType()
 			return s
 		}(),
@@ -194,7 +192,7 @@ func TestPingSourceStatusIsReady(t *testing.T) {
 			s.InitializeConditions()
 			s.MarkSchedule()
 			s.MarkSink(exampleUri)
-			s.PropagateDeploymentAvailability(sources.TestHelper.AvailableDeployment())
+			s.PropagateDeploymentAvailability(availableDeployment)
 			s.PropagateDeploymentAvailability(&appsv1.Deployment{})
 			return s
 		}(),
@@ -207,7 +205,7 @@ func TestPingSourceStatusIsReady(t *testing.T) {
 			s.InitializeConditions()
 			s.MarkSchedule()
 			s.MarkSink(exampleUri)
-			s.PropagateDeploymentAvailability(sources.TestHelper.AvailableDeployment())
+			s.PropagateDeploymentAvailability(availableDeployment)
 			s.MarkNoEventType("Testing", "")
 			return s
 		}(),
@@ -220,7 +218,7 @@ func TestPingSourceStatusIsReady(t *testing.T) {
 			s.InitializeConditions()
 			s.MarkSchedule()
 			s.MarkSink(nil)
-			s.PropagateDeploymentAvailability(sources.TestHelper.AvailableDeployment())
+			s.PropagateDeploymentAvailability(availableDeployment)
 			return s
 		}(),
 		wantConditionStatus: corev1.ConditionFalse,
@@ -232,7 +230,7 @@ func TestPingSourceStatusIsReady(t *testing.T) {
 			s.InitializeConditions()
 			s.MarkSchedule()
 			s.MarkSink(nil)
-			s.PropagateDeploymentAvailability(sources.TestHelper.AvailableDeployment())
+			s.PropagateDeploymentAvailability(availableDeployment)
 			s.MarkSink(exampleUri)
 			return s
 		}(),
@@ -283,7 +281,7 @@ func TestPingSourceStatusGetTopLevelCondition(t *testing.T) {
 		s: func() *PingSourceStatus {
 			s := &PingSourceStatus{}
 			s.InitializeConditions()
-			s.PropagateDeploymentAvailability(sources.TestHelper.AvailableDeployment())
+			s.PropagateDeploymentAvailability(availableDeployment)
 			return s
 		}(),
 		want: &apis.Condition{
@@ -332,7 +330,7 @@ func TestPingSourceStatusGetTopLevelCondition(t *testing.T) {
 			s := &PingSourceStatus{}
 			s.InitializeConditions()
 			s.MarkSink(exampleUri)
-			s.PropagateDeploymentAvailability(sources.TestHelper.AvailableDeployment())
+			s.PropagateDeploymentAvailability(availableDeployment)
 			return s
 		}(),
 		want: &apis.Condition{
@@ -359,7 +357,7 @@ func TestPingSourceStatusGetTopLevelCondition(t *testing.T) {
 			s.InitializeConditions()
 			s.MarkSchedule()
 			s.MarkSink(exampleUri)
-			s.PropagateDeploymentAvailability(sources.TestHelper.AvailableDeployment())
+			s.PropagateDeploymentAvailability(availableDeployment)
 			return s
 		}(),
 		want: &apis.Condition{
@@ -373,7 +371,7 @@ func TestPingSourceStatusGetTopLevelCondition(t *testing.T) {
 			s.InitializeConditions()
 			s.MarkSchedule()
 			s.MarkSink(exampleUri)
-			s.PropagateDeploymentAvailability(sources.TestHelper.AvailableDeployment())
+			s.PropagateDeploymentAvailability(availableDeployment)
 			s.MarkEventType()
 			return s
 		}(),
@@ -388,7 +386,7 @@ func TestPingSourceStatusGetTopLevelCondition(t *testing.T) {
 			s.InitializeConditions()
 			s.MarkSchedule()
 			s.MarkSink(exampleUri)
-			s.PropagateDeploymentAvailability(sources.TestHelper.AvailableDeployment())
+			s.PropagateDeploymentAvailability(availableDeployment)
 			s.PropagateDeploymentAvailability(&appsv1.Deployment{})
 			return s
 		}(),
@@ -405,7 +403,7 @@ func TestPingSourceStatusGetTopLevelCondition(t *testing.T) {
 			s.InitializeConditions()
 			s.MarkSchedule()
 			s.MarkSink(exampleUri)
-			s.PropagateDeploymentAvailability(sources.TestHelper.AvailableDeployment())
+			s.PropagateDeploymentAvailability(availableDeployment)
 			s.MarkNoEventType("Testing", "")
 			return s
 		}(),
@@ -420,7 +418,7 @@ func TestPingSourceStatusGetTopLevelCondition(t *testing.T) {
 			s.InitializeConditions()
 			s.MarkSchedule()
 			s.MarkSink(nil)
-			s.PropagateDeploymentAvailability(sources.TestHelper.AvailableDeployment())
+			s.PropagateDeploymentAvailability(availableDeployment)
 			return s
 		}(),
 		want: &apis.Condition{
@@ -436,7 +434,7 @@ func TestPingSourceStatusGetTopLevelCondition(t *testing.T) {
 			s.InitializeConditions()
 			s.MarkSchedule()
 			s.MarkSink(nil)
-			s.PropagateDeploymentAvailability(sources.TestHelper.AvailableDeployment())
+			s.PropagateDeploymentAvailability(availableDeployment)
 			s.MarkSink(exampleUri)
 			return s
 		}(),
@@ -487,7 +485,7 @@ func TestPingSourceStatusGetCondition(t *testing.T) {
 		s: func() *PingSourceStatus {
 			s := &PingSourceStatus{}
 			s.InitializeConditions()
-			s.PropagateDeploymentAvailability(sources.TestHelper.AvailableDeployment())
+			s.PropagateDeploymentAvailability(availableDeployment)
 			return s
 		}(),
 		condQuery: PingSourceConditionReady,
@@ -528,7 +526,7 @@ func TestPingSourceStatusGetCondition(t *testing.T) {
 			s.InitializeConditions()
 			s.MarkSchedule()
 			s.MarkSink(exampleUri)
-			s.PropagateDeploymentAvailability(sources.TestHelper.AvailableDeployment())
+			s.PropagateDeploymentAvailability(availableDeployment)
 			return s
 		}(),
 		condQuery: PingSourceConditionReady,
@@ -543,7 +541,7 @@ func TestPingSourceStatusGetCondition(t *testing.T) {
 			s.InitializeConditions()
 			s.MarkSchedule()
 			s.MarkSink(exampleUri)
-			s.PropagateDeploymentAvailability(sources.TestHelper.AvailableDeployment())
+			s.PropagateDeploymentAvailability(availableDeployment)
 			s.MarkEventType()
 			return s
 		}(),
@@ -559,7 +557,7 @@ func TestPingSourceStatusGetCondition(t *testing.T) {
 			s.InitializeConditions()
 			s.MarkSchedule()
 			s.MarkSink(exampleUri)
-			s.PropagateDeploymentAvailability(sources.TestHelper.AvailableDeployment())
+			s.PropagateDeploymentAvailability(availableDeployment)
 			s.MarkNoSink("Testing", "hi%s", "")
 			return s
 		}(),
@@ -577,7 +575,7 @@ func TestPingSourceStatusGetCondition(t *testing.T) {
 			s.InitializeConditions()
 			s.MarkSchedule()
 			s.MarkSink(exampleUri)
-			s.PropagateDeploymentAvailability(sources.TestHelper.AvailableDeployment())
+			s.PropagateDeploymentAvailability(availableDeployment)
 			s.MarkInvalidSchedule("Testing", "hi%s", "")
 			return s
 		}(),
@@ -595,7 +593,7 @@ func TestPingSourceStatusGetCondition(t *testing.T) {
 			s.InitializeConditions()
 			s.MarkSchedule()
 			s.MarkSink(exampleUri)
-			s.PropagateDeploymentAvailability(sources.TestHelper.AvailableDeployment())
+			s.PropagateDeploymentAvailability(availableDeployment)
 			s.PropagateDeploymentAvailability(&appsv1.Deployment{})
 			return s
 		}(),
@@ -613,7 +611,7 @@ func TestPingSourceStatusGetCondition(t *testing.T) {
 			s.InitializeConditions()
 			s.MarkSchedule()
 			s.MarkSink(exampleUri)
-			s.PropagateDeploymentAvailability(sources.TestHelper.AvailableDeployment())
+			s.PropagateDeploymentAvailability(availableDeployment)
 			s.PropagateDeploymentAvailability(&appsv1.Deployment{})
 			return s
 		}(),
@@ -631,7 +629,7 @@ func TestPingSourceStatusGetCondition(t *testing.T) {
 			s.InitializeConditions()
 			s.MarkSchedule()
 			s.MarkSink(exampleUri)
-			s.PropagateDeploymentAvailability(sources.TestHelper.AvailableDeployment())
+			s.PropagateDeploymentAvailability(availableDeployment)
 			s.MarkEventType()
 			s.MarkNoEventType("Testing", "hi")
 			return s
@@ -648,7 +646,7 @@ func TestPingSourceStatusGetCondition(t *testing.T) {
 			s.InitializeConditions()
 			s.MarkSchedule()
 			s.MarkSink(nil)
-			s.PropagateDeploymentAvailability(sources.TestHelper.AvailableDeployment())
+			s.PropagateDeploymentAvailability(availableDeployment)
 			return s
 		}(),
 		condQuery: PingSourceConditionReady,
@@ -665,7 +663,7 @@ func TestPingSourceStatusGetCondition(t *testing.T) {
 			s.InitializeConditions()
 			s.MarkSchedule()
 			s.MarkSink(nil)
-			s.PropagateDeploymentAvailability(sources.TestHelper.AvailableDeployment())
+			s.PropagateDeploymentAvailability(availableDeployment)
 			s.MarkSink(exampleUri)
 			return s
 		}(),
