+++ pkg/apis/sources/v1beta1/apiserver_conversion_test.go
@@ -18,10 +18,9 @@ package v1beta1
 
 import (
 	"context"
+	"errors"
 	"reflect"
 	"testing"
-
-	"knative.dev/eventing/pkg/apis/sources"
 
 	"github.com/google/go-cmp/cmp"
 	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
@@ -30,24 +29,18 @@ import (
 	duckv1 "knative.dev/pkg/apis/duck/v1"
 )
 
-/*
-Copyright 2020 The Knative Authors
+type dummyObject struct{}
 
-Licensed under the Apache License, Version 2.0 (the "License");
-you may not use this file except in compliance with the License.
-You may obtain a copy of the License at
+func (*dummyObject) ConvertTo(ctx context.Context, obj apis.Convertible) error {
+	return errors.New("Won't go")
+}
 
-    http://www.apache.org/licenses/LICENSE-2.0
-
-Unless required by applicable law or agreed to in writing, software
-distributed under the License is distributed on an "AS IS" BASIS,
-WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-See the License for the specific language governing permissions and
-limitations under the License.
-*/
+func (*dummyObject) ConvertFrom(ctx context.Context, obj apis.Convertible) error {
+	return errors.New("Won't go")
+}
 
 func TestApiServerSourceConversionBadType(t *testing.T) {
-	good, bad := &ApiServerSource{}, &sources.DummyObject{}
+	good, bad := &ApiServerSource{}, &dummyObject{}
 
 	if err := good.ConvertTo(context.Background(), bad); err == nil {
 		t.Errorf("ConvertTo() = %#v, wanted error", bad)
