+++ pkg/apis/sources/v1beta1/apiserver_types.go
@@ -48,32 +48,6 @@ var (
 	_ duckv1.KRShaped    = (*ApiServerSource)(nil)
 )
 
-const (
-	// ApiServerSourceAddEventType is the ApiServerSource CloudEvent type for adds.
-	ApiServerSourceAddEventType = "dev.knative.apiserver.resource.add"
-	// ApiServerSourceUpdateEventType is the ApiServerSource CloudEvent type for updates.
-	ApiServerSourceUpdateEventType = "dev.knative.apiserver.resource.update"
-	// ApiServerSourceDeleteEventType is the ApiServerSource CloudEvent type for deletions.
-	ApiServerSourceDeleteEventType = "dev.knative.apiserver.resource.delete"
-
-	// ApiServerSourceAddRefEventType is the ApiServerSource CloudEvent type for ref adds.
-	ApiServerSourceAddRefEventType = "dev.knative.apiserver.ref.add"
-	// ApiServerSourceUpdateRefEventType is the ApiServerSource CloudEvent type for ref updates.
-	ApiServerSourceUpdateRefEventType = "dev.knative.apiserver.ref.update"
-	// ApiServerSourceDeleteRefEventType is the ApiServerSource CloudEvent type for ref deletions.
-	ApiServerSourceDeleteRefEventType = "dev.knative.apiserver.ref.delete"
-)
-
-// ApiServerSourceEventTypes is the list of CloudEvent types the ApiServerSource emits.
-var ApiServerSourceEventTypes = []string{
-	ApiServerSourceAddEventType,
-	ApiServerSourceDeleteEventType,
-	ApiServerSourceUpdateEventType,
-	ApiServerSourceAddRefEventType,
-	ApiServerSourceDeleteRefEventType,
-	ApiServerSourceUpdateRefEventType,
-}
-
 // ApiServerSourceSpec defines the desired state of ApiServerSource
 type ApiServerSourceSpec struct {
 	// inherits duck/v1 SourceSpec, which currently provides:
